api_requirements.txt
uvicorn
FastAPI
SpeechRecognition
phonemizer
# backend: sudo apt-get install espeak
uvicorn
fastapi
PyYAML
gtts
pydub
# sudo apt-get install ffmpeg
py3langid
setuptools
python-dotenv
google-generativeai
python-multipart
torch
torchaudio
openai
transformers